Remove plone.app.stagingbehavior
The version of plone.app.iterate that ships with Plone 5 (v3.1.1) provides iterate support out of the box to all dexterity content types by default.  There is no longer any need for this package to be used, and in fact at start up p.a.iterate warns of impending doom when p.a.stagingbehavior is installed.